Move addSettingsToGroup to configuration module
